Primeiro build

+++ Backend/Gestao-Acolhidos/src/main/java/org/felipe/gestaoacolhidos/model/domain/dto/Hosted/ReferenceAddress/ReferenceAddressDTO.java
@@ -6,6 +6,6 @@ public record ReferenceAddressDTO(
         String neighborhood,
         int number,
         String cep,
-        int phoneNumber
+        long phoneNumber
 ) {
 }
